+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## TBD
+## [2.2.2] 2017-12-01
 ### Fixed
 * Metastore status check now works for tunneled connections. #issue-34